+++ cmd/forklift/dev/plt/pallets.go
@@ -164,7 +164,7 @@ func checkAction(versions Versions) cli.ActionFunc {
 			return err
 		}
 
-		if _, _, err := fcli.CheckPallet(0, pallet, cache); err != nil {
+		if _, _, err := fcli.Check(0, pallet, cache); err != nil {
 			return err
 		}
 		return nil
@@ -186,7 +186,7 @@ func planAction(versions Versions) cli.ActionFunc {
 			return err
 		}
 
-		if _, _, err = fcli.PlanPallet(0, pallet, cache, c.Bool("parallel")); err != nil {
+		if _, _, err = fcli.Plan(0, pallet, cache, c.Bool("parallel")); err != nil {
 			return err
 		}
 		return nil
@@ -216,7 +216,7 @@ func stageAction(versions Versions) cli.ActionFunc {
 		if err != nil {
 			return err
 		}
-		if err = fcli.StagePallet(pallet, stageStore, cache, versions.NewBundle); err != nil {
+		if _, err = fcli.StagePallet(pallet, stageStore, cache, versions.NewBundle); err != nil {
 			return err
 		}
 		fmt.Println("Done! To apply the staged pallet, you can run `sudo -E forklift stage apply`.")
@@ -228,21 +228,26 @@ func stageAction(versions Versions) cli.ActionFunc {
 
 func applyAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, true, true)
+		pallet, repoCache, err := processFullBaseArgs(c, true, true)
 		if err != nil {
 			return err
 		}
 		if err = fcli.CheckCompatibility(
-			pallet, cache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
 			c.Bool("ignore-tool-version"),
 		); err != nil {
 			return err
 		}
-
-		if err := fcli.ApplyPallet(0, pallet, cache, c.Bool("parallel")); err != nil {
+		workspace, err := forklift.LoadWorkspace(c.String("workspace"))
+		if err != nil {
 			return err
 		}
-		fmt.Println()
+
+		if err = fcli.ApplyPallet(
+			pallet, repoCache, workspace, versions.NewStageStore, versions.NewBundle, c.Bool("parallel"),
+		); err != nil {
+			return err
+		}
 		fmt.Println("Done!")
 		return nil
 	}