+++ cmd/forklift/plt/pallets.go
@@ -101,27 +101,28 @@ func switchAction(versions Versions) cli.ActionFunc {
 		// version is incompatible
 
 		// cache everything required by pallet
-		pallet, cache, err := processFullBaseArgs(c, false)
+		pallet, repoCache, err := processFullBaseArgs(c, false)
 		if err != nil {
 			return err
 		}
 		if err = fcli.CheckShallowCompatibility(
-			pallet, cache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
 			c.Bool("ignore-tool-version"),
 		); err != nil {
 			return err
 		}
 		if _, err = fcli.CacheAllRequirements(
-			pallet, cache.Path(), cache, c.Bool("include-disabled"), c.Bool("parallel"),
+			pallet, repoCache.Path(), repoCache, c.Bool("include-disabled"), c.Bool("parallel"),
 		); err != nil {
 			return err
 		}
 
 		// apply pallet
-		if err = fcli.ApplyPallet(0, pallet, cache, c.Bool("parallel")); err != nil {
-			return errors.Wrap(err, "couldn't deploy local pallet")
+		if err = fcli.ApplyPallet(
+			pallet, repoCache, workspace, versions.NewStageStore, versions.NewBundle, c.Bool("parallel"),
+		); err != nil {
+			return err
 		}
-		fmt.Println()
 		fmt.Println("Done!")
 		return nil
 	}
@@ -248,7 +249,7 @@ func checkAction(versions Versions) cli.ActionFunc {
 			return err
 		}
 
-		if _, _, err := fcli.CheckPallet(0, pallet, cache); err != nil {
+		if _, _, err := fcli.Check(0, pallet, cache); err != nil {
 			return err
 		}
 		return nil
@@ -270,7 +271,7 @@ func planAction(versions Versions) cli.ActionFunc {
 			return err
 		}
 
-		if _, _, err = fcli.PlanPallet(0, pallet, cache, c.Bool("parallel")); err != nil {
+		if _, _, err = fcli.Plan(0, pallet, cache, c.Bool("parallel")); err != nil {
 			return errors.Wrap(
 				err, "couldn't deploy local pallet (have you run `forklift plt cache` recently?)",
 			)
@@ -302,7 +303,7 @@ func stageAction(versions Versions) cli.ActionFunc {
 		if err != nil {
 			return err
 		}
-		if err = fcli.StagePallet(pallet, stageStore, cache, versions.NewBundle); err != nil {
+		if _, err = fcli.StagePallet(pallet, stageStore, cache, versions.NewBundle); err != nil {
 			return err
 		}
 		fmt.Println("Done! To apply the staged pallet, you can run `sudo -E forklift stage apply`.")
@@ -314,23 +315,26 @@ func stageAction(versions Versions) cli.ActionFunc {
 
 func applyAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, true)
+		pallet, repoCache, err := processFullBaseArgs(c, true)
 		if err != nil {
 			return err
 		}
 		if err = fcli.CheckCompatibility(
-			pallet, cache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
 			c.Bool("ignore-tool-version"),
 		); err != nil {
 			return err
 		}
-
-		if err := fcli.ApplyPallet(0, pallet, cache, c.Bool("parallel")); err != nil {
-			return errors.Wrap(
-				err, "couldn't deploy local pallet (have you run `forklift plt cache` recently?)",
-			)
+		workspace, err := forklift.LoadWorkspace(c.String("workspace"))
+		if err != nil {
+			return err
 		}
-		fmt.Println()
+
+		if err = fcli.ApplyPallet(
+			pallet, repoCache, workspace, versions.NewStageStore, versions.NewBundle, c.Bool("parallel"),
+		); err != nil {
+			return err
+		}
 		fmt.Println("Done!")
 		return nil
 	}