+++ cmd/forklift/stage/bundles.go
@@ -2,11 +2,14 @@ package stage
 
 import (
 	"fmt"
+	"slices"
 	"strconv"
+	"strings"
 
 	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
 
+	"github.com/PlanktoScope/forklift/internal/app/forklift"
 	fcli "github.com/PlanktoScope/forklift/internal/app/forklift/cli"
 )
 
@@ -22,29 +25,59 @@ func lsBunAction(versions Versions) cli.ActionFunc {
 			return errMissingStore
 		}
 
+		names := getBundleNames(store)
 		indices, err := store.List()
 		if err != nil {
 			return err
 		}
 		for _, index := range indices {
-			bundle, err := store.LoadFSBundle(index)
-			if err != nil {
-				fmt.Printf("%d: Error: couldn't load bundle: %s\n", index, err)
-				continue
-			}
-			fmt.Printf("%d: %s@%s", index, bundle.Def.Pallet.Path, bundle.Def.Pallet.Version)
-			if !bundle.Def.Pallet.Clean {
-				fmt.Print(" (staged with uncommitted pallet changes)")
-			}
-			if bundle.Def.Includes.HasOverrides() {
-				fmt.Print(" (staged with overridden pallet requirements)")
-			}
-			// TODO: add label for the last successfully-applied bundle, and the next one staged to be
-			// deployed, i.e. the pending apply (if it exists), and the rollback bundle (if it exists)
-			fmt.Println()
+			printBundleSummary(store, index, names)
 		}
 		return nil
 	}
+}
+
+func getBundleNames(store *forklift.FSStageStore) map[int][]string {
+	names := make(map[int][]string)
+	for name, index := range store.Def.Stages.Names {
+		names[index] = append(names[index], name)
+	}
+	for _, indexNames := range names {
+		slices.Sort(indexNames)
+	}
+	if index, ok := store.GetRollback(); ok {
+		names[index] = slices.Concat([]string{"rollback"}, names[index])
+	}
+	if index, ok := store.GetNext(); ok {
+		names[index] = slices.Concat([]string{"next"}, names[index])
+	}
+	if index, ok := store.GetCurrent(); ok {
+		names[index] = slices.Concat([]string{"current"}, names[index])
+	}
+	if index, ok := store.GetPending(); ok {
+		names[index] = slices.Concat([]string{"pending"}, names[index])
+	}
+	return names
+}
+
+func printBundleSummary(store *forklift.FSStageStore, index int, names map[int][]string) {
+	bundle, err := store.LoadFSBundle(index)
+	if err != nil {
+		fmt.Printf("%d: Error: couldn't load bundle: %s\n", index, err)
+		return
+	}
+	fmt.Print(index)
+	if indexNames := names[index]; len(indexNames) > 0 {
+		fmt.Printf(" (%s)", strings.Join(indexNames, ", "))
+	}
+	fmt.Printf(": %s@%s", bundle.Def.Pallet.Path, bundle.Def.Pallet.Version)
+	if !bundle.Def.Pallet.Clean {
+		fmt.Print(" (staged with uncommitted pallet changes)")
+	}
+	if bundle.Def.Includes.HasOverrides() {
+		fmt.Print(" (staged with overridden pallet requirements)")
+	}
+	fmt.Println()
 }
 
 // show-bun
@@ -62,13 +95,13 @@ func showBunAction(versions Versions) cli.ActionFunc {
 		rawIndex := c.Args().First()
 		index, err := strconv.Atoi(rawIndex)
 		if err != nil {
-			return errors.Wrapf(err, "Couldn't parse staged bundle index %s as an integer", rawIndex)
+			return errors.Wrapf(err, "couldn't parse staged bundle index %s as an integer", rawIndex)
 		}
 		bundle, err := store.LoadFSBundle(index)
 		if err != nil {
 			return errors.Wrapf(err, "couldn't load staged bundle %d", index)
 		}
-		fcli.PrintStagedBundle(0, store, bundle, index)
+		fcli.PrintStagedBundle(0, store, bundle, index, getBundleNames(store)[index])
 		return nil
 	}
 }