+++ cmd/forklift/stage/store.go
@@ -1,23 +1,345 @@
 package stage
 
 import (
+	"fmt"
+	"strconv"
+
 	"github.com/pkg/errors"
+	"github.com/urfave/cli/v2"
 
 	"github.com/PlanktoScope/forklift/internal/app/forklift"
+	fcli "github.com/PlanktoScope/forklift/internal/app/forklift/cli"
 )
 
 var errMissingStore = errors.Errorf(
-	"you first need to stage a pallet, e.g. with `forklift plt stage`",
+	"no pallets have been staged yet: you first must stage a pallet, e.g. with `forklift plt stage`",
 )
+
+func loadNextBundle(
+	wpath string, versions Versions,
+) (*forklift.FSBundle, *forklift.FSStageStore, error) {
+	store, err := getStageStore(wpath, versions)
+	if err != nil {
+		return nil, nil, err
+	}
+	if !store.Exists() {
+		return nil, store, errMissingStore
+	}
+
+	index, ok := store.GetNext()
+	if !ok {
+		return nil, store, errors.Errorf(
+			"no next staged pallet bundle to apply: you first must set a pallet to stage next, " +
+				"e.g. with `forklift stage set-next`",
+		)
+	}
+	if store.NextFailed() {
+		fmt.Printf("Next stage failed in the past, so we won't use it: %d\n", index)
+		if index, ok = store.GetCurrent(); !ok {
+			return nil, store, errors.Errorf(
+				"the next staged pallet bundle already failed, and no staged pallet bundle was " +
+					"applied successfully in the past, so we have no fallback!",
+			)
+		}
+		fmt.Printf("Current stage, as a fallback: %d\n", index)
+	} else {
+		if pending, ok := store.GetPending(); ok && index == pending {
+			fmt.Printf("Next stage is pending: %d\n", index)
+		} else if current, ok := store.GetCurrent(); ok && index == current {
+			fmt.Printf("Next stage previously had a successful apply: %d\n", index)
+		} else {
+			fmt.Printf("Next stage: %d\n", index)
+		}
+	}
+
+	bundle, err := store.LoadFSBundle(index)
+	if err != nil {
+		return nil, store, errors.Wrapf(err, "couldn't load staged pallet bundle %d", index)
+	}
+	return bundle, store, nil
+}
 
 func getStageStore(wpath string, versions Versions) (*forklift.FSStageStore, error) {
 	workspace, err := forklift.LoadWorkspace(wpath)
 	if err != nil {
 		return nil, err
 	}
-	cache, err := workspace.GetStageStore(versions.NewStageStore)
+	store, err := workspace.GetStageStore(versions.NewStageStore)
 	if err != nil {
 		return nil, err
 	}
-	return cache, nil
+	return store, nil
+}
+
+func checkShallowCompatibility(
+	bundle *forklift.FSBundle, versions Versions, ignoreTool bool,
+) error {
+	if ignoreTool {
+		fmt.Printf(
+			"Warning: ignoring the tool's version (%s) for version compatibility checking!\n",
+			versions.Tool,
+		)
+	}
+
+	if err := fcli.CheckArtifactCompatibility(
+		bundle.Def.ForkliftVersion, versions.Tool, versions.MinSupportedBundle, bundle.Path(),
+		ignoreTool,
+	); err != nil {
+		return errors.Wrapf(
+			err, "forklift tool has a version incompatibility with staged pallet bundle %s",
+			bundle.Path(),
+		)
+	}
+	return nil
+}
+
+// show
+
+func showAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		store, err := getStageStore(c.String("workspace"), versions)
+		if err != nil {
+			return err
+		}
+		if !store.Exists() {
+			return errMissingStore
+		}
+
+		indent := 0
+		fcli.IndentedPrintf(indent, "Stage store %s:\n", store.Path())
+		indent++
+		names := getBundleNames(store)
+
+		// TODO: display info about whether the swapfile exists
+		fcli.IndentedPrint(indent, "Next staged pallet bundle to be applied:")
+		next, hasNext := store.GetNext()
+		if !hasNext {
+			fmt.Println(" (none)")
+		} else {
+			fmt.Printf(" %d\n", next)
+			printNextSummary(indent+1, store, next, names[next])
+		}
+
+		fcli.IndentedPrint(indent, "Last successfully-applied staged pallet bundle:")
+		current, hasCurrent := store.GetCurrent()
+		switch {
+		case !hasCurrent:
+			fmt.Println(" (none)")
+		case next == current:
+			fmt.Printf(" %d (see above)\n", current)
+		default:
+			fmt.Printf(" %d\n", current)
+			printCurrentSummary(indent+1, store, current, names[current])
+		}
+
+		fcli.IndentedPrint(indent, "Previous successfully-applied staged pallet bundle:")
+		rollback, hasRollback := store.GetRollback()
+		if !hasRollback {
+			fmt.Println(" (none)")
+		} else {
+			fmt.Printf(" %d\n", rollback)
+			printRollbackSummary(indent+1, store, rollback, names[rollback])
+		}
+
+		return nil
+	}
+}
+
+func printNextSummary(indent int, store *forklift.FSStageStore, index int, names []string) {
+	bundle, err := store.LoadFSBundle(index)
+	if err != nil {
+		fmt.Printf("Error: couldn't load staged bundle d (was it deleted?): %s\n", err.Error())
+		return
+	}
+
+	printBasicSummary(indent, bundle, names)
+	failed := store.NextFailed()
+	pending, hasPending := store.GetPending()
+	isPending := (hasPending && index == pending)
+	current, hasCurrent := store.GetCurrent()
+	isCurrent := (hasCurrent && index == current)
+	if failed || isPending || isCurrent {
+		fcli.IndentedPrint(indent, "Status: ")
+		switch {
+		case failed && hasCurrent:
+			fmt.Printf(
+				"failed to be applied; the last successfully-applied staged pallet bundle (%d) will be "+
+					"used instead\n",
+				current,
+			)
+		case failed:
+			fmt.Println(
+				"failed to be applied; no pallet will be applied until another pallet is staged",
+			)
+		case isPending:
+			fmt.Println("not yet applied; will be used for the next apply")
+		case hasCurrent && index == current:
+			fmt.Println("already applied; will still be used for the next apply")
+		}
+	}
+}
+
+func printBasicSummary(indent int, bundle *forklift.FSBundle, names []string) {
+	fcli.IndentedPrint(indent, "Staged names:")
+	if len(names) == 0 {
+		fmt.Println(" (none)")
+	} else {
+		fmt.Println()
+		for _, name := range names {
+			fcli.BulletedPrintln(indent+1, name)
+		}
+	}
+
+	fcli.IndentedPrintf(indent, "Pallet: %s@%s\n", bundle.Def.Pallet.Path, bundle.Def.Pallet.Version)
+
+	indent++
+	if !bundle.Def.Pallet.Clean {
+		fcli.BulletedPrintln(indent, "Staged with uncommitted pallet changes")
+	}
+	if bundle.Def.Includes.HasOverrides() {
+		fcli.BulletedPrint(indent, "Staged with overridden pallet requirements")
+	}
+}
+
+func printCurrentSummary(indent int, store *forklift.FSStageStore, index int, names []string) {
+	bundle, err := store.LoadFSBundle(index)
+	if err != nil {
+		fmt.Printf("Error: couldn't load staged bundle %d (was it deleted?): %s\n", index, err.Error())
+		return
+	}
+
+	printBasicSummary(indent, bundle, names)
+}
+
+func printRollbackSummary(indent int, store *forklift.FSStageStore, index int, names []string) {
+	bundle, err := store.LoadFSBundle(index)
+	if err != nil {
+		fmt.Printf("Error: couldn't load staged bundle %d (was it deleted?): %s\n", index, err.Error())
+		return
+	}
+
+	printBasicSummary(indent, bundle, names)
+}
+
+// show-hist
+
+func showHistAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		store, err := getStageStore(c.String("workspace"), versions)
+		if err != nil {
+			return err
+		}
+		if !store.Exists() {
+			return errMissingStore
+		}
+
+		names := getBundleNames(store)
+		for _, index := range store.Def.Stages.History {
+			printBundleSummary(store, index, names)
+		}
+		if index, ok := store.GetPending(); ok {
+			printBundleSummary(store, index, names)
+		}
+		return nil
+	}
+}
+
+// set-next
+
+func setNextAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		store, err := getStageStore(c.String("workspace"), versions)
+		if err != nil {
+			return err
+		}
+		if !store.Exists() {
+			return errMissingStore
+		}
+
+		rawNewNext := c.Args().First()
+		newNext, err := strconv.Atoi(rawNewNext)
+		if err != nil {
+			return errors.Wrapf(err, "couldn't parse staged bundle index %s as an integer", rawNewNext)
+		}
+		if _, err = store.LoadFSBundle(newNext); err != nil {
+			return errors.Wrapf(err, "couldn't load staged bundle %d", newNext)
+		}
+
+		if next, hasNext := store.GetNext(); hasNext {
+			fmt.Printf("Changing the next staged pallet bundle from %d to %d...\n", next, newNext)
+		} else {
+			fmt.Printf("Setting the next staged pallet bundle to %d...\n", newNext)
+		}
+		store.SetNext(newNext)
+		if err = store.CommitState(); err != nil {
+			return errors.Wrap(err, "couldn't commit updated stage store state")
+		}
+		fmt.Println("Done!")
+		return nil
+	}
+}
+
+// check
+
+func checkAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		bundle, _, err := loadNextBundle(c.String("workspace"), versions)
+		if err != nil {
+			return err
+		}
+		if err = checkShallowCompatibility(
+			bundle, versions, c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+		if _, _, err = fcli.Check(0, bundle, bundle); err != nil {
+			return err
+		}
+		return nil
+	}
+}
+
+// plan
+
+func planAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		bundle, _, err := loadNextBundle(c.String("workspace"), versions)
+		if err != nil {
+			return err
+		}
+		if err = checkShallowCompatibility(
+			bundle, versions, c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+		if _, _, err = fcli.Plan(0, bundle, bundle, c.Bool("parallel")); err != nil {
+			return err
+		}
+		return nil
+	}
+}
+
+// apply
+
+func applyAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		bundle, store, err := loadNextBundle(c.String("workspace"), versions)
+		if err != nil {
+			return errors.Wrap(
+				err, "couldn't load last successfully-applied pallet bundle as a fallback",
+			)
+		}
+		if err = checkShallowCompatibility(
+			bundle, versions, c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+		fmt.Println()
+
+		if err = fcli.ApplyNextOrCurrentBundle(0, store, bundle, c.Bool("parallel")); err != nil {
+			return err
+		}
+		fmt.Println("Done!")
+		return nil
+	}
 }